refactor(StatsSummary): extract to separate file
